Assume that any unspecified party/committee memberships starts today, not at the beginning of the session.
As discussed with Stortinget, party or committee relations can change also mid-session. Since Stortinget's
API only provides a snapshot of the current state (in the dagensrepresentanter resource), we must assume
that the membership starts today (which will work just fine with our daily imports).

This may lead to slightly wrong dates if we miss a daily import, which we'll then need to adjust by hand.